# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cinema API",
        "description": "API for Cinema",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Localhost API Server"
        }
    ],
    "paths": {
...
        "/api/salles": {
            "get": {
                "tags": [
                    "Salles"
                ],
                "summary": "Liste toutes les salles",
                "description": "Récupère la liste de toutes les salles de cinéma.",
                "operationId": "e7188100fab9b75ceed273ede574c261",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nom": {
                                                "type": "string",
                                                "example": "Salle 1"
                                            },
                                            "capacite": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "Normale"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Salles"
                ],
                "summary": "Créer une nouvelle salle",
                "description": "Ajoute une nouvelle salle et génère automatiquement les sièges.",
                "operationId": "d3b1a2743044328cece5e0b41a81fdd0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "capacite",
                                    "type"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Salle 1"
                                    },
                                    "capacite": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "Normale",
                                            "VIP"
                                        ],
                                        "example": "VIP"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Salle créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "salle": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nom": {
                                                    "type": "string",
                                                    "example": "Salle 1"
                                                },
                                                "capacite": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "VIP"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/salles/{id}": {
            "get": {
                "tags": [
                    "Salles"
                ],
                "summary": "Récupérer une salle",
                "operationId": "a287b158cafff514dd9c53ee7a49f674",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la salle",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la salle",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Salles"
                ],
                "summary": "Modifier une salle existante",
                "operationId": "793e23dc6c0e57920025ac90b8b9ed0c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la salle",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Salle mise à jour avec succès"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Salles"
                ],
                "summary": "Supprimer une salle",
                "operationId": "c33650c54b56a6994dc255fba62d9e12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la salle",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Salle supprimée avec succès"
                    }
                }
            }
        },
...
    },
...
}